+++ libImaging/RankFilter.c
@@ -61,8 +61,8 @@ ImagingRankFilter(Imaging im, int size, int rank)
         return (Imaging) ImagingError_ValueError("bad filter size");
     
     /* malloc check ok, for overflow in the define below */
-    if (size > SIZE_MAX / size || 
-        size > SIZE_MAX / (size * sizeof(FLOAT32))) {
+    if (size > INT_MAX / size || 
+        size > INT_MAX / (size * sizeof(FLOAT32))) {
         return (Imaging) ImagingError_ValueError("filter size too large");
     }
